Move data fields to common
Moves response fields that are used across packages to common.
