+++ service/common/data.go
@@ -1,0 +1,35 @@
+// Copyright 2022 Upbound Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package common
+
+import "github.com/google/uuid"
+
+// RelationshipSet represents set of relationships.
+type RelationshipSet map[string]interface{}
+
+// AttributeSet represents set of attributes.
+type AttributeSet map[string]interface{}
+
+// Meta represents metadata.
+type Meta map[string]interface{}
+
+// DataSet represents a set of data for an entity.
+type DataSet struct {
+	Type            string    `json:"type"`
+	ID              uuid.UUID `json:"id"`
+	AttributeSet    `json:"attributes"`
+	RelationshipSet `json:"relationships"`
+	Meta            `json:"meta"`
+}
